provisioning-portal/requirements.txt
flask
docker
argon2-cffi
pyyaml
gunicorn
gevent
gevent-websocket
flask-socketio